Move setAuthorizationRoleForm to bh
