+++ src/views/repository/Settings.vue
@@ -58,7 +58,7 @@ import RepositoryViewBase from '@/components/repository/RepositoryViewBase';
 import RepositoryBase from './Base';
 import EditProfileForm from '@/components-v1/user/EditProfileForm';
 import EditRepositoryForm from '@/components-v1/repository/EditRepositoryForm';
-import SetAuthorizationRoleForm from '@/components-v1/repository/SetAuthorizationRoleForm';
+import SetAuthorizationRoleForm from '@/components/repository/SetAuthorizationRoleForm';
 import AuthorizationsList from '@/components-v1/repository/AuthorizationsList';
 import AuthorizationRequestsList from '@/components-v1/repository/AuthorizationRequestsList';
 import LoginForm from '@/components-v1/auth/LoginForm';
